[Breaking Change] (test) Refactor the testing data and settings as 2 modules to let other testings to use.
