+++ test/unit_test/_config.py
@@ -1,0 +1,13 @@
+from pathlib import Path
+
+
+Test_CSV_File_Path: str = str(Path("./for_testing.csv"))
+Test_XLSX_File_Path: str = str(Path("./for_testing.xlsx"))
+Test_JSON_File_Path: str = str(Path("./for_testing.json"))
+Test_XML_File_Path: str = str(Path("./for_testing.xml"))
+Test_PROPERTIES_File_Path: str = str(Path("./for_testing.properties"))
+
+Test_Writing_Mode: str = "a+"
+Test_XML_Writing_Mode: str = "wb"
+Test_Reading_Mode: str = "r"
+
